Create Webscraper and Server
Add webscraper to retrieve data on Animal Crossing critters from wikis and export the data to JSON or text. The webscraper will export dates and time as a string in the JSON export which isn't easy to work with.

Add a server with minimal API functionality. The server will only return all bugs in JSON currently.

Add a gitignore file to keep the repo clean, optimized, and prevent sensitive data from being seen.