Add Cookies, Refactor Layout, and Refactor Code
Add cookies to the webpage to ensure that the application remembers which critters a user has donated and start the user on the critter tab they interacted with last.

Refactor the layout of the page to use a vertical navbar that looks more similar to discord to freeup some vertical realestate for the critter table.

Refactor the code to move critter filtering logic into its own javascript file to make the codebase more modular and easier to maintain. Fix all runtime exceptions shown in the browser console. Remove code that connects to the critter api and instead use json files local to the frontend to populate critters. Since the critter data is static, this will not reduce functionality.